rm ml workers completely for further test with worker threads
